GH-9999: Fix `FluxMessageChannel` for restarted application
Fixes: #9999
Issue link: https://github.com/spring-projects/spring-integration/issues/9999

After the application context stop, the `FluxMessageChannel` continues to try to emit messages.
Just because subscriptions to the provided publishers are not cancelled.
More over, we would like to come back to the production from those publishers when we start
application back.

* Fix `FluxMessageChannel` implementing a `Lifecycle` contract.
Gather provided publisher in a local cache to come back to them when we call `start()`,
essentially, initiating a new subscription to those provided publishers

**Auto-cherry-pick to `6.3.x`**

# Conflicts:
#	spring-integration-core/src/main/java/org/springframework/integration/channel/AbstractMessageChannel.java