Update SI Core to add Nullability to module
This is a first pass at the module to make sure it is being converted properly.

Related to https://github.com/spring-projects/spring-integration/issues/10083

Add nullability to org.springframework.messaging.core package

Add nullability to org.springframework.aggregator & org.springframework.acks packages

Note in this commit that there cases where methods are called where the parameters passed in may contain a null.
In those cases the attribute that contains the potential null is tested with an `Assert.notNull`.  This may not be the best approach, let's discuss.

Also note that in FluxAggregatorMessageHandler line 112 `@NullUnmarked` is used on the the applyWindowOptions method.
This is because the consumer.apply on line 121 can not be null according to JSpecify.   But we can't make that assumption for how the user implemented the function.
There maybe a better way of handling this besides marking the method as `NullUnmarked`.   Let's discuss.

Similarly in FluxAggregatorMessageHandler the way in which sequenceSizeHeader is utilized it had to be `@NullUnmarked`.   Let's discuss

Make sure all annotations are applied consistently

Apply code changes from code review

Fix nullability in the `FluxAggregatorMessageHandler`

* This PR is not ready for merge.   There are failing tests, but wanted to capture where it is at.
* Also still having issue with the .apply() in FluxAggregatorMessageHandler
* Make `windowSizeFunction` as `Function<Message<?>, @Nullable Integer>`
because `sequenceSizeHeader()` may return `null` from message headers
* Extract local `subscriptionToDispose` in the `stop()` to satisfy null check context
* Use `Objects.requireNonNull(signal.get())` to satisfy `Function.apply()` contract.
The `if (signal.hasValue()) {` does the trick for us, but currently that is not visible
for that `signal.get()`
* Remove `@NullUnmarked` since we have just mitigated all the null problems

Updated tests so that they work with nullability changes

Update the tests so that they will pass with nullify changes

Resolve second set of test failures

* Remove Nullunmarked from FluxAggregatorMessageHandler
* Add TaskScheuler to tests that are failing because of nullability
* In the past the TaskScheduler attribute in the testscould be set to null and the tests would succeed.
But because of nullability additions these need to be populated

Add TaskScheduler for adapter in testNet tests

Update code based on second review.

* Added style to IDE to resolve style issues
* Resolved the missed remove from IntegrationContextUtils.
* Made sure that private test methods were marked static
* Updated Assert.NotNulls to Assert.state and updated their messages
* Rebased and updated nullification for changes

Add jspecify Nullability checks to SI

* Update formating errors from previous commits
* Remove ignore SuppressWarning flags for methods that have been updated to be null safe
* Make sure added private methods in tests are static
* Update tests so they will succeed after the addition of nullification
* Remove Disable from tests
* Add appropriate beans, BeanFactories to tests that now fail because of nullify
* Remove Mock and use createTestApplicationContext
* Apply fixes recommended by Artem
* Create environment such that the tests have the beans necessary to attain their test target
* Update the TestApplicationContainerAware so that it can handle multiple context refreshes
  Since gradle can run multiple tests across multiple threads the static context can be refreshed multiple times.
  Thus we need to add code to handle the potential exception
* You will note that are few classes where mocking is used instead of the components of TestApplicationContextAware.
  This is because test relied on mocking for additional setup of the test or checks

Update code to reflect requested changes made during code review

* Rename CONTEXT to TEST_INTEGRATION_CONTEXT for increased clarity of purpose.
* Remove Duplicate bean registrations from tests that are already handled by TestUtils.createTestApplicationContext
* Replace @BeforAll and AfterAll prep and cleanup with the TestApplicationContextAware interface in PayloadAndHeaderMappingTests.java
* Set BeanFactory for ThreadAffinityClientConnectionFactory.java
* Add setBeanFactory method to JmsInboundGateway.java so that users can set its beanfactory as sell as for its endpoint
* Added javadocs to the TestApplicationContextAware interface.
* Make sure that ExpressionEvaluatingCorrelationStrategy.getCorrelationKey returns non null
* Rebased

Update ExpressionEvaluatingCorrelationStrategy.getCorrelationKey nullable

Remove SupressWarnings from IntegrationObjectSupport.afterPropertiesSet

Optimize CorrelatingMessageBarrier so that methods only check for nullability on the onInit