Replace sample classes in CompositeCodecTests with records to reduce maintenance
Rename field names of foo and foo2 in CompositeCodecTests to professional descriptive names
In the onlyDefaultCodec implementation in the CompositeCodecTests use a class other than a PojoCodec class to simulate a real world case.
Enhance javadoc for CompositeCodec to include information on ClassUtils.findClosestMatch and its effects on codec selection.
Make sure @see is in the correct location in javadoc