Fix race condition for `ApplicationEventListeningMessageProducer`
The `ContextStoppedEvent` and `ContextClosedEvent` are emitted when the
application context is not in an active state anymore.
Therefore, dispatching those events into beans (`MessageChannel` in our case)
is not what possible to handle because endpoints are stopped (or destroyed)

* Fix `ApplicationEventListeningMessageProducer` to ignore `ContextStoppedEvent` and `ContextClosedEvent`
altogether for output channel of `AbstractMessageChannel`.
The warning we emit in that case should be enough to indicate what is going on and what could be done
to mitigate situation with custom `MessageChannel` implementation.
* Improve error message for the `MessageDispatchingException` in the `AbstractMessageChannel.assertApplicationRunning()`
to indicate what exactly channel is in the problem.