Fix `SpringIntegrationTestExecutionListener` for restart
Related to: https://github.com/spring-projects/spring-framework/issues/35168

Spring Framework has introduced recently an optimization for cached contexts.
So, if one is not used, it is stopped in the cache.
When we pull it next time from there, it is restarted.
The `SpringIntegrationTestExecutionListener` ha a flaw to change the state of the endpoints
in the application context, but never reset it back to original.
The problem has become apparent because of not reset auto-startup status of endpoints and that `AC.restart()`

* Fix `SpringIntegrationTestExecutionListener` to use `setAutoStartup(false)`
on `autoStartupCandidates` in the `afterTestClass()`.
This is exactly a state the `MockIntegrationContext` has gathered them in its initialization phase.
* Fix `MockIntegrationContext.autoStartupCandidates` to be a `Set` instead.
The point is that `postProcessBeforeInitialization()` deals with `AbstractEndpoint` beans.
And then `afterSingletonsInstantiated()` also gathers those beans from the `BeanFactory`.
So, with a `Set` we aim for no duplications