CompositeCodec constructors require at least one delegate
* Adjust the behavior of the class so that delegates are required.

* In the current implementation, users could pass in an empty list to the composite codec.
* This is an CompositeCodec and if we had no codecs, then why have the class.
* All @Nullable's were removed from the delegate attribute.
* Null returns from the findDelegate have been removed.   The default delegate will be returned in the stead.
* Update the tests to require at least one codec in the delegates

* Update code to meet code review requirements

* Add Author tag to affected classes

* Add tests to the CompositeCodecTests to tests delegates

And also test if a user attempts to an invalid class.

* Update deprecated ctor so that it will only establish default codec

* Simplify the initialization of the `codec` attribute

* Update the javadoc for the CompositeCodec class.

Change the  `ClassUtils.findClosestMatch` `failOnTie` to true.
  In the past when finding the appropriate delegate(codec) for a object and
  their were multiple delegates available the first one was returned.
  The problem is it could return a different delegate on each launch of the `findDelegate`.
  Now with `failOnTie` set to true a `IllegalStateException` is returned.

Remove deprecated constructor from CompositeCodec

* Replace sample classes in CompositeCodecTests with records to reduce maintenance

Rename field names of foo and foo2 in CompositeCodecTests to professional descriptive names
In the onlyDefaultCodec implementation in the CompositeCodecTests use a class other than a PojoCodec class to simulate a real world case.
Enhance javadoc for CompositeCodec to include information on ClassUtils.findClosestMatch and its effects on codec selection.
Make sure @see is in the correct location in javadoc

* Add CompositeCodec documentation to reference documentation

Provide example for using CompositeCodec.
Remove @See in CompositeCodec javadocs.   This is handled in the links in the javadoc

* In codec.adoc all sentences must be on same line

Add code snippet into docs for classes
Verify no tabs exist in sample