Fix Kotlin DSL according to IDEA errors and suggestions (#10307)
* Fix Kotlin DSL according to IDEA errors and suggestions

It is not clear why it does not fail with Gradle, but right now it is not possible to build the project using IDEA compiler.

Some of the suggestions are good and have led to the Nullability fixes in related Java DSL classes.
The `Any` extension requirement for the generic argument of reified types is not clear, but works.
The requirement to refine nullable Kotlin types is good.

Somewhat related to: https://github.com/spring-projects/spring-integration/issues/10083

* * Address more spotted Nullability concerns

* More Nullability refinement in the `BaseIntegrationFlowDefinition`