GH-10314: Revise `TcpListener.onMessage()` contract for `void` (#10315)
Fixes: https://github.com/spring-projects/spring-integration/issues/10314

Looks like a `boolean` return for the `TcpListener.onMessage()` is a leftover
of some earlier design or some idea which didn't make it into the project.
On the other hand, all the logic in the project around this `onMessage()` usage
is properly handled by the delegation or exceptions.

* Change `TcpListener.onMessage()` to have a `void` as return type
* Fix all the respective usages and simplify some implementations with removing those bogus `return false;`
and using proper `if..else` logic if necessary
* Fix some code style in the affected classes, like proper `assertThat` or diamond operators