GH-10083: Apply Nullability to the `transformer` package
Related to: https://github.com/spring-projects/spring-integration/issues/10083

The `transformer` pattern cannot produce `null` at all.
Therefore, making respective `Transformer.transform()` contract as non-null
leads to changes in many places where it was possible to return null
with subsequent `ReplyRequiredException` thrown from the `handler`.

The current fix is a bit radical and transformer-specific:
whenever we encounter `null` in the `transform()` call, we thrown a
`MessageTransformationException` instead.
Which is indeed making more sense from the `Transformer` perspective.
We may treat it as a breaking change, but the docs still state clearly that there is an exception in case of null.
Although it is different now.

The Nullability work here has also led to the review of the `getComponentType()` implementation
for the `Transformer` hierarchy.
Plus, the `AbstractMessageProcessingTransformer` is now an `IntegrationObjectSupport`.
The `MessageTransformingHandler` performs respective propagation in its `doInit()` to to the `transformer`
if that one is `IntegrationObjectSupport`.

The `HeaderEnricher` doesn't do a hard reference to the provided `Map` anymore to avoid
potential mutation of the source externally.
This has led to the change in the `HeaderEnricherSpec` to create a `HeaderEnricher` later in the `doGet()`
with all respective properties propagation.

Some of the test have suffered changes as well according to a new logic of the `transformer` package