GH-10372: Fix the ignoring logic for tmp files
Fixes: https://github.com/spring-projects/spring-integration/issues/10372

The `AbstractInboundFileSynchronizingMessageSource` uses a `CompositeFileListFilter`
for a combination of the user-provided filter and then the one to ignore temporary files.
However, the `CompositeFileListFilter` logic is `OR`, which means the temporary
file might be accepted by the user-provided.

* Fix the `AbstractInboundFileSynchronizingMessageSource.buildFilter()`
to use a `ChainFileListFilter` instead (with an `AND` logic), and put `ignoreTemporaryFiles`
filter as the first one.
* Fix (S)FTP docs for an actual logic regarding local files filtering.

**Cherry-pick to `6.5.x`**