GH-10362: Fix `PostgresSubscribableChannel` for race condition
Fixes: https://github.com/spring-projects/spring-integration/issues/10362

Unsubscribing message handlers while processing an in-flight message may lead to a `MessageDeliveryException`.
There is a possibility in the `PostgresSubscribableChannel` to have a polled message dispatched when no subscribers to the channel anymore.
The `doPollAndDispatchMessage()` can enter to the `pollMessage()` when it sees `hasHandlers`.
However, the last subscriber to the channel is unsubscribed in parallel.

* Fix `PostgresSubscribableChannel` using a `ReadWriteLock` to block `unsubscribe()` until currently polled message is processed.
Or opposite: do not let to `doPollAndDispatchMessage()` to enter `pollMessage()` logic until `unsubscribe()` is done.
This way the `hasHandlers` state is guarded by lock synchronization.

Signed-off-by: Norbert Schneider <mail@bertschneider.de>

**Auto-cherry-pick to `6.4.x`**
# Conflicts:
#	spring-integration-jdbc/src/main/java/org/springframework/integration/jdbc/channel/PostgresSubscribableChannel.java