Use `isNoop()` instead of `Observation.NOOP`
The Micrometer team is considering some improvements in the Observation API which involves changing the behavior of `Observation.NOOP`: micrometer-metrics/micrometer#6700
The `Observation.NOOP` is not truly no-op right now, it does context propagation, we are trying to make it truly no-op in the PR above. 
The non-truly no-op version will still be used in certain scenarios (see `NoopButScopeHandlingObservation`). 
The `observation.isNoop()` should be used for checking if an `Observation` is no-op or not instead of `Observation.NOOP`.