GH-10431: Propagate observation to error handler
Fixes: https://github.com/spring-projects/spring-integration/issues/10431

The `SourcePollingChannelAdapter` is able to start an observation for a just received message from a source.
This observation is completed when the polling task is done (including transaction).
However, a thrown from downstream error goes to the `errorChannel` already after finishing the mentioned
observation (even if that includes error).
According to end-user experience, the error handling must be a part of the same trace started for the received message.

* Fix `AbstractPollingEndpoint` to perform `donePollingTask()` only for a successful message reception
* Fix `SourcePollingChannelAdapter` to propagate its `ObservationRegistry` to the `ErrorHandlingTaskExecutor`
* Introduce an `Observation` handling in the `ErrorHandlingTaskExecutor`.
Essentially, stop currently on scope `Observation` after error handling for th running task failure
* Prove the feature with a new `SourcePollingChannelAdapterErrorObservationTests`

**Cherry-pick to `6.5.x`**