Kafka MessageDrivenAdapterTests: batch race condition
The `ConsumerConfig.FETCH_MIN_BYTES_CONFIG` is `1` byte by default.
Therefore, the batch consumer expectations might be fulfilled too early
just for one record.
Adjust such a config to `12` bytes to be sure that both published records
are received in a single batch: exactly what we expect in the test logic.

**Auto-cherry-pick to `6.4.x`**