GH-10090: Add `AmqpClientInboundGateway` (#10447)
* GH-10090: Add `AmqpClientInboundGateway`

Related to: https://github.com/spring-projects/spring-integration/issues/10090

* Add `AmqpClientInboundGateway` that is mostly a copy/paste of the `AmqpClientMessageProducer`,
but adds a reply-producing logic
* Cover with tests and document this new component
* Fix a couple typos in the `amqp-1.0.adoc`

* Fix docs typos

* Extract the `IntegrationRabbitAmqpMessageListener` as a top-level (package private) class.
Serves as a reusable unit of work for both `AmqpClientMessageProducer` & `AmqpClientInboundGateway`.
The one-way and request-reply parts are handled as a `BiConsumer` action injection
into this `IntegrationRabbitAmqpMessageListener` instance.
Now both `AmqpClientMessageProducer` & `AmqpClientInboundGateway` are much simpler.
* Improve `AmqpClientMessageProducer.processRequest` Javadoc
mentioning that `requestMessage` param is out of use
* Be more specific with the type for `UnsupportedOperationException`
in the `IntegrationRabbitAmqpMessageListener`