Update gradle to 9.2.0 (#10537)
* Update gradle to 9.2.0

Attempted to resolve TODO on line 1099, however some of the API is
still not recognized.  Added a 9.2+ to the TODO

* Remove Gradle workaround from schema processing

The TODO comment indicated this workaround was needed until Gradle
9.1.1 or 9.2. Since Gradle has been updated to 9.2.0 in the
workaround is no longer necessary.

This change simplifies the schema loading code by using the more
concise `?.withInputStream` pattern instead of the verbose null-check
and conditional logic that was required by earlier Gradle versions.