Add `nullChannel()` to Kotlin DSL
This commit adds support for `nullChannel()` in the Kotlin DSL,
mirroring the functionality recently added to the Groovy DSL.
The `nullChannel()` method populates a NullChannel instance at the
current IntegrationFlow chain position, acting like "/dev/null" to
discard messages.

Two test cases verify the functionality:
- Messages sent to a flow ending with `nullChannel()` are discarded
- Transform operations can precede `nullChannel()` in a flow

This enhancement provides feature parity between the Kotlin and
Groovy DSLs for null channel support.

* Remove nullChannelFLow test

* nullChannelWithTransformFlow is satisfactory
* Simplify transformer declaration for the transformer null test
* Replace import <>.* with proper package names
* Remove unused id's in all transforms used by tests
* Remove commented code.  Users can find this in the docs.

* Add back id's to tests

* ReAdd commented code set and description
* Update DsL for transform

* Fix formatting of documentation

* Add proper spacing around comment

This is so that it is associated with the proper code

* Remove space on empty line